+++ src/main/java/team/cheese/controller/sale/SaleController.java
@@ -117,29 +117,29 @@ public class SaleController {
         return "/sale/saleManage";
     }
 
-    // 수정하기 버튼을 눌렀을 때 글을 받아서 jsp로 전달
-    @PostMapping("/modify")
-    public String modify(@RequestParam Long no, Model model, HttpServletRequest request) throws Exception {
-
-        Map map = saleService.modify(no);
-        SaleDto saleDto = (SaleDto) map.get("saleDto");
-        String tagContents = (String) map.get("tagContents");
-        HttpSession session = request.getSession();
-        String user_id = (String) session.getAttribute("userId");
-        String user_nick = (String) session.getAttribute("userNick");
-
-        saleDto.setSeller_id(user_id);
-        saleDto.setSeller_nick(user_nick);
-
-        List<ImgDto> imglist = imgService.read(saleDto.getGroup_no());
-
-        model.addAttribute("Sale", saleDto);
-        model.addAttribute("Tag", tagContents);
-        model.addAttribute("imglist", imglist); // model로 값 전달
-        model.addAttribute("saleCategory1", saleCategoryDao.selectCategory1());
-
-        return "/sale/saleWrite";
-    }
+//    // 수정하기 버튼을 눌렀을 때 글을 받아서 jsp로 전달
+//    @GetMapping("/modify")
+//    public String modify(@RequestParam Long no, Model model, HttpServletRequest request) throws Exception {
+//
+//        Map map = saleService.modify(no);
+//        SaleDto saleDto = (SaleDto) map.get("saleDto");
+//        String tagContents = (String) map.get("tagContents");
+//        HttpSession session = request.getSession();
+//        String user_id = (String) session.getAttribute("userId");
+//        String user_nick = (String) session.getAttribute("userNick");
+//
+//        saleDto.setSeller_id(user_id);
+//        saleDto.setSeller_nick(user_nick);
+//
+//        List<ImgDto> imglist = imgService.read(saleDto.getGroup_no());
+//
+//        model.addAttribute("Sale", saleDto);
+//        model.addAttribute("Tag", tagContents);
+//        model.addAttribute("imglist", imglist); // model로 값 전달
+//        model.addAttribute("saleCategory1", saleCategoryDao.selectCategory1());
+//
+//        return "/sale/saleWrite";
+//    }
 
     @RequestMapping("/remove")
     public String remove(@RequestParam Long no, Model model, HttpSession session) throws Exception {
